requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
av==15.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cmake==4.1.3
coacd==1.0.7
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==4.1.1
decorator==5.2.1
deepdiff==8.6.1
Deprecated==1.3.1
deprecation==2.1.0
diffusers==0.35.2
dill==0.4.0
distro==1.9.0
docopt==0.6.2
draccus==0.10.0
einops==0.8.2
etils==1.13.0
faiss-cpu==1.13.2
Farama-Notifications==0.0.4
fast_simplification==0.1.13
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.9.0
genesis-world==0.4.0
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
groq==1.0.0
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
inquirerpy==0.3.4
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
lancedb==0.29.2
langchain-core==1.2.14
langsmith==0.7.6
lazy_loader==0.4
lerobot==0.4.3
libigl==2.6.1
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
mergedeep==1.3.4
moviepy==2.2.1
mpmath==1.3.0
mujoco==3.5.0
multidict==6.7.1
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.6.1
num2words==0.5.14
numba==0.64.0
numpy==2.2.6
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.12.0.88
OpenEXR==3.4.6
orderly-set==5.5.0
orjson==3.11.7
packaging==25.0
pandas==3.0.1
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.9.2
plotly==5.24.1
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pycollada==0.9.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGEL3D==0.6.1
pyglet==2.1.13
pygltflib==1.16.0
Pygments==2.19.2
pymeshlab==2025.7.post1
pynput==1.8.1
pyobjc-core==12.1
pyobjc-framework-ApplicationServices==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-CoreText==12.1
pyobjc-framework-Quartz==12.1
PyOpenGL==3.1.10
pyparsing==3.3.2
pyserial==3.5
pysplashsurf==0.14.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyyaml-include==1.4.1
quadrants==0.4.0
regex==2026.2.19
requests==2.32.5
requests-toolbelt==1.0.0
rerun-sdk==0.26.2
rich==14.3.3
rtree==1.4.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
sentry-sdk==2.53.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.5
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
tetgen==0.8.2
threadpoolctl==3.6.0
tifffile==2026.3.3
tokenizers==0.22.2
toml==0.10.2
torch==2.7.1
torchcodec==0.5
torchvision==0.22.1
tqdm==4.67.3
transformers==4.57.6
trimesh==4.11.2
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
vtk==9.6.0
wandb==0.24.2
wcwidth==0.6.0
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
z3-solver==4.15.4.0
zipp==3.23.0
zstandard==0.25.0
